// src/services/models/ccUsers.ts
// Generated by https://quicktype.io

export interface ccUsers {
  Id: number;
  Title: string;
  PERSON_RANK: string;
  PERSON_ACTIVE: string;
  PERSON_DIRECTORATE: string;
  PERSON_EMAIL: string;
  PERSON_ROLE: string;
  P_LAST_NAME: string;
  P_FIRST_NAME: string;
  PERSON_ATTRIBUTES: string;
  PERSON_TRAINING: string;
  ID: number;
  Created: string;
}